Fix issue project isn't shown sometime in report table
This is a WORKAROUND which uses synchronous call to get project list from server.

This will fix the issue that project isn't shown on report table.
This is because the response for GET-entries might arrive earlier than GET-projects
even though GET-projects request is issued firstly. In this case, the report table
will be filled while the project list is not available yet.